refactor(remote): Factor out InstantTypeAdapter and DurationTypeAdapter
